Add bumplimit control
Previously it was possible to create more than 50 threads for each board
which is not something that's supposed to happen.

The correct behavior is removing the topic, that has the oldest last
post in it and adding a brand new topic to the top of the list.

I've implemented that as an after_create callback inside Topic model.