Fix pagination
For some reason pagination was not working properly, so I've replaced
will_paginate with Kaminari.

When that didn't help, I realized it's the wrong scope that was causing
all sorts of trouble.

The goal was to display 10 topics per page, 5 pages total. The most
recently commented topic should be on top. If it's a new topic with no
posts in it, it should be on top as well.

It was implemented by using "touch: true" on the Topic model and sorting
the output by updated_at attribute.