Refactor the 500 post per topic spec
Previously it was creating 500 objects inside a feature spec which was
very slow and not a good way to test this.

Instead, we now have a topic model spec that make sure the bumplimit?
method functions properly, as well as a controller spec that makes sure
the before_filter gets used on create, which utilizes the bumplimit?
method on topic to make sure more than 500 posts per topic won't be
created.