Rename thread to topic
Previously, there was some inconsistency across the app where in certain
places the resource was called a "thread" while in others it was called
a "topic".

Originally, the classic naming convention for imageboards is "thread".
However, Rails reserves the word "thread" as a special one. Because of
that I decided to rename the resource across the app.